chore: add view function to replace state variabel currentPrice
The currentPrice variable is not being used from with the contract.
It is more gas efficient to move the calculation of the current
price to a function which can be called from the contract instead
of storing the value.

The code got refactored to reflect this change. The actual calculation
got moved to an private function which now is used by the new function
calculateCurrentPrice() as well as the function calculateTotalCost()