feat: add TDF transfer exception for Treasury
- Why
20% of TDF tokens should be minted as $Sweat reserve. These tokens
should be minted before the Sale goes live to ensure the price
on the bonding curve is accurate. After minting the tokens to the
reserve, they have to be transferrable to the contributors
- How
By adding an exception for the transfer, i.e. if the transfer is
initiated through the treasury it will not revert while when
a token buyer or holder would like to transfer TDF tokens it
would revert